fix: 코멘트 로직 분리
