+++ week07/Shawn/src/hooks/queries/useGetInfiniteLpComment.ts
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from "@tanstack/react-query"
-import { getLpComments } from "../../apis/lp"
+import { getLpComments } from "../../apis/lpComment"
 import { QUERY_KEY } from "../../constant/key"
 
 function useGetInfiniteLpComment(lpId: number, limit: number, search: string, order: 'asc' | 'desc') {
